Reworked secret server responses

+++ bin/secret-server/main.py
@@ -11,16 +11,14 @@ from fastapi.responses import PlainTextResponse
 if not os.path.isdir("data"):
     os.mkdir("data")
 
-db = sqlite3.connect("data/db.sqlite3", check_same_thread=False)
-if db.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='secrets'").fetchone() is None:
-    db.execute("CREATE TABLE secrets (key TEXT PRIMARY KEY, value TEXT)")
-    db.commit()
+db = sqlite3.connect("data/db.sqlite3", autocommit=True, check_same_thread=False)
+db.execute("CREATE TABLE IF NOT EXISTS secrets (key TEXT PRIMARY KEY, value TEXT)")
 
 app = FastAPI(docs_url=None, redoc_url=None)
 
 
 @app.middleware("http")
-async def check_auth(request: Request, call_next):
+async def check_auth(request: Request, next):
     auth = request.headers.get("Authorization")
     if not auth:
         return PlainTextResponse("Unauthorized", status_code=401, headers={"WWW-Authenticate": "Basic"})
@@ -33,7 +31,7 @@ async def check_auth(request: Request, call_next):
     if not secrets.compare_digest(password, os.environ.get("SECRET_SERVER_PASSWORD", "admin")):
         return PlainTextResponse("Unauthorized", status_code=401, headers={"WWW-Authenticate": "Basic"})
 
-    return await call_next(request)
+    return await next(request)
 
 
 @app.get("/")
@@ -41,21 +39,25 @@ def get_root():
     return {key: value for key, value in db.execute("SELECT key, value FROM secrets")}
 
 
-@app.get("/{key}", response_class=PlainTextResponse)
+@app.get("/{key}")
 def get_key(key: str):
     value = db.execute("SELECT value FROM secrets WHERE key = ?", (key,)).fetchone()
-    return value[0] if value else ""
+    if not value:
+        return PlainTextResponse("", status_code=404)
+
+    return PlainTextResponse(value[0])
 
 
-@app.post("/{key}", response_class=PlainTextResponse)
+@app.post("/{key}")
 async def post_key(key: str, request: Request):
     db.execute("INSERT OR REPLACE INTO secrets (key, value) VALUES (?, ?)", (key, await request.body()))
-    db.commit()
-    return "OK"
+    return PlainTextResponse("OK")
 
 
-@app.delete("/{key}", response_class=PlainTextResponse)
+@app.delete("/{key}")
 def delete_key(key: str):
+    if not db.execute("SELECT 1 FROM secrets WHERE key = ?", (key,)).fetchone():
+        return PlainTextResponse("Not Found", status_code=404)
+
     db.execute("DELETE FROM secrets WHERE key = ?", (key,))
-    db.commit()
-    return "OK"
+    return PlainTextResponse("OK")
